nil/copy: Add a simpler CopyGOBLines trait
This lets us implement all the complex logic of copying a GOB in a
central place and all we really need for each GOB type is an iterator
function.  There's a bit of duplication here with the copy functions but
it's still better than what we had before.

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/35660>