panvk: Prepare panvk_draw_prepare_varyings for JM indirect
On JM hardware, we need to allocate a buffer depending on vertex count.

As a result, for indirect and indexed draw we allocate a large buffer with
alloc on fault set.

The size of that buffer is calculated assuming a max of 2 millions vertices
and 18 attributes per vertex (16 user attributes, 2 specials)

Signed-off-by: Mary Guillemard <mary.guillemard@collabora.com>
Reviewed-by: Olivia Lee <olivia.lee@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/35724>