egl/x11: Allow the loader to choose zink
If the loader tells us to use zink, we use zink.  This breaks the
Nouveau heuristic where we use Zink instead of the Nouveau GL driver on
Turing and later.  The only reason why it works at all is because we
fail to load what the loader requests and then EGL goes "Oops. No
driver! Let's see if Zink works" and we pick it up on the fallback.  We
shouldn't depend on those fallbacks if the loader explicitly says to use
Zink for some bit of hardware.

Reviewed-by: Adam Jackson <ajax@redhat.com>
Reviewed-By: Mike Blumenkrantz <michael.blumenkrantz@gmail.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/36014>