ci: Build weston 14.0.1 from source
There is a bug in Weston 10 that's causes instability when we don't have
wl_drm which isn't likely to get fixed in a point release.  Most of CI
is fine but the final patch in this MR causes AMD raven to kill weston
part-way through runs, destroying the run.  Just update weston to
14.0.1.

Reviewed-by: Valentine Burley <valentine.burley@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/36026>