egl/x11: Rework LIBGL_KOPPER_DRI2 checks
This moves the multibuffers check earlier to make it clear that this is
an early-exit.  We also delete kopper_without_multibuffers in favor of a
local variable that doesn't leak all over everywhere.  The new variable
is named force_zink because that's really what we're doing: force-enable
Zink even if the X11 checks fail.

Reviewed-by: Adam Jackson <ajax@redhat.com>
Reviewed-By: Mike Blumenkrantz <michael.blumenkrantz@gmail.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/36014>