mesa/st: Re-instate test for shader_clock cap
With the check for the cap gone no driver will expose
ARB_shader_clock. Unfortunately the CI doesn't catch this
because it doesn't provide expectations whether a test should
pass or be skipped. In this case

  spec@arb_shader_clock@execution@clock
  spec@arb_shader_clock@execution@clock2x32

went from pass to skip. (Tested on r600, but on radeonsi one
can also see that the extension ARB_shader_clock is no longer
available).

Adding the test for the cap back in fixes this.

Fixes: 2ce201707e2a00 (Add support for EXT_shader_clock)

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/35992>